backend/requirements.txt
fastapi==0.104.0
uvicorn==0.22.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pillow==10.0.1
pytest==7.4.2
scikit-learn==1.3.1
python-dotenv==1.0.0
email-validator==2.0.0
bcrypt==4.0.1

